ProductRecommender/requirements.txt
# Core data processing libraries
pandas>=1.3.0
numpy>=1.20.0

# Machine learning libraries
scikit-learn>=1.0.0
gensim>=4.0.0

# For HTML parsing in description_processor.py
beautifulsoup4>=4.9.0
lxml>=4.6.0

# Natural language processing
nltk>=3.6.0

# For visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# For saving/loading ML models
joblib>=1.0.0

# For dimension reduction visualization
scikit-learn>=1.0.0 

# For color processing
colormath>=3.0.0

# Database operations
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Environment variables management
python-dotenv>=0.19.0

# HTTP requests
requests>=2.26.0
